# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API",
        "description": "Opentact API",
        "version": "1.0.0"
    },
    "host": "portal.opentact.org",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "paths": {
...
        "/api/balance": {
            "get": {
                "tags": [
                    "payments"
                ],
                "summary": "Return user balance",
                "parameters": [
                    {
                        "name": "userid",
                        "in": "query",
                        "description": "APP User ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Balance"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/addCredit": {
            "post": {
                "tags": [
                    "payments"
                ],
                "summary": "Add user's credit",
                "parameters": [
                    {
                        "name": "userid",
                        "in": "formData",
                        "description": "APP User ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "amount",
                        "in": "formData",
                        "description": "Amount",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "remark",
                        "in": "formData",
                        "description": "Remark",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success result"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/creditHistory": {
            "get": {
                "tags": [
                    "payments"
                ],
                "summary": "Get credit history",
                "parameters": [
                    {
                        "name": "userid",
                        "in": "query",
                        "description": "APP User ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Credit history"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/getAllowedCountry": {
            "get": {
                "tags": [
                    "payments"
                ],
                "summary": "Return allowed country",
                "parameters": [
                    {
                        "name": "userid",
                        "in": "query",
                        "description": "APP User ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Allowed country"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/getRates": {
            "get": {
                "tags": [
                    "payments"
                ],
                "summary": "Return rates for exact country",
                "parameters": [
                    {
                        "name": "userid",
                        "in": "query",
                        "description": "APP User ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Country",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rates"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/getRate": {
            "get": {
                "tags": [
                    "payments"
                ],
                "summary": "Return rates for exact number",
                "parameters": [
                    {
                        "name": "userid",
                        "in": "query",
                        "description": "APP User ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "number",
                        "in": "query",
                        "description": "Number",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rates"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/app/create": {
            "post": {
                "tags": [
                    "app"
                ],
                "summary": "Create new APP",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "App name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "account_id",
                        "in": "formData",
                        "description": "Developer Account ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "sign",
                        "in": "formData",
                        "description": "Auth sign",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success result"
                    },
                    "401": {
                        "description": "Auth required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}